+++ app/packages/app/src/components/FlashlightLooker/FlashlightLooker.module.css
@@ -1,0 +1,6 @@
+.flashlightLooker {
+  width: 100%;
+  height: 100%;
+  display: block;
+  position: relative;
+}
