+++ app/packages/app/src/components/FlashlightLooker/hooks.ts
@@ -1,0 +1,19 @@
+import { useRecoilTransaction_UNSTABLE } from "recoil";
+import { modal, selectedSamples } from "../../recoil/atoms";
+import { useEventHandler } from "../../utils/hooks";
+
+export const useEscape = () => {
+  useEventHandler(
+    document,
+    "keydown",
+    useRecoilTransaction_UNSTABLE(
+      ({ get, set }) =>
+        (event: KeyboardEvent) => {
+          event.key === "Escape" &&
+            !get(modal) &&
+            set(selectedSamples, new Set());
+        },
+      []
+    )
+  );
+};
