+++ app/packages/app/src/components/Group/index.ts
@@ -1,0 +1,1 @@
+export { default as Group } from "./Group";
