+++ app/packages/app/src/components/Sidebar/Entries/EntryCounts.tsx
@@ -3,10 +3,10 @@ import { selectorFamily } from "recoil";
 
 import * as aggregationAtoms from "../../../recoil/aggregations";
 import { matchedTags } from "../../../recoil/filters";
+import { MATCH_LABEL_TAGS } from "../../../recoil/sidebar";
 import { State } from "../../../recoil/types";
 
 import { SuspenseEntryCounts } from "../../Common/CountSubcount";
-import { MATCH_LABEL_TAGS } from "../utils";
 
 export const PathEntryCounts = ({
   modal,
@@ -38,14 +38,16 @@ const labelTagCount = selectorFamily<
   { modal: boolean; tag: string; extended: boolean }
 >({
   key: `labelTagCount`,
-  get: ({ tag, ...rest }) => ({ get }) =>
-    get(
-      aggregationAtoms.cumulativeCounts({
-        path: tag,
-        ...MATCH_LABEL_TAGS,
-        ...rest,
-      })
-    )[tag] || 0,
+  get:
+    ({ tag, ...rest }) =>
+    ({ get }) =>
+      get(
+        aggregationAtoms.cumulativeCounts({
+          path: tag,
+          ...MATCH_LABEL_TAGS,
+          ...rest,
+        })
+      )[tag] || 0,
 });
 
 export const tagIsMatched = selectorFamily<
@@ -53,19 +55,23 @@ export const tagIsMatched = selectorFamily<
   { key: State.TagKey; tag: string; modal: boolean }
 >({
   key: "tagIsActive",
-  get: ({ key, tag, modal }) => ({ get }) =>
-    get(matchedTags({ key, modal })).has(tag),
-  set: ({ key, tag, modal }) => ({ get, set }, toggle) => {
-    const atom = matchedTags({ key, modal });
-    const current = get(atom);
+  get:
+    ({ key, tag, modal }) =>
+    ({ get }) =>
+      get(matchedTags({ key, modal })).has(tag),
+  set:
+    ({ key, tag, modal }) =>
+    ({ get, set }, toggle) => {
+      const atom = matchedTags({ key, modal });
+      const current = get(atom);
 
-    set(
-      atom,
-      toggle
-        ? new Set([tag, ...current])
-        : new Set([...current].filter((t) => t !== tag))
-    );
-  },
+      set(
+        atom,
+        toggle
+          ? new Set([tag, ...current])
+          : new Set([...current].filter((t) => t !== tag))
+      );
+    },
 });
 
 export const LabelTagCounts = ({
