+++ app/packages/app/src/components/Sidebar/Entries/FilterEntry.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useDebounce } from "react-use";
 import { useRecoilState } from "recoil";
-import { textFilter } from "../recoil";
+import { textFilter } from "../../../recoil/sidebar";
 import { InputDiv } from "./utils";
 
 const Filter = ({ modal }: { modal: boolean }) => {
