+++ app/packages/app/src/components/Sidebar/index.tsx
@@ -1,7 +1,4 @@
 export * as Entries from "./Entries";
-export * from "./utils";
-export * from "./recoil";
-
 import Sidebar from "./Sidebar";
 
 export default Sidebar;
