+++ app/packages/components/src/components/ColumnScroller/ColumnScroller.tsx
@@ -27,7 +27,17 @@ const ColumnScroller: React.FC<ColumnScrollerProps> = ({
     flashlight.attach(id);
   }, [flashlight, id]);
 
-  return <div id={id}></div>;
+  return (
+    <div
+      style={{
+        display: "block",
+        width: "100%",
+        height: "100%",
+        position: "relative",
+      }}
+      id={id}
+    ></div>
+  );
 };
 
 export default ColumnScroller;
